Added Strider, Zoglin, Zombified Piglin. Removed MCE Mobs
Added the remaining 1.16 mobs to the unsupported section of the plugin, as well as a 1.15 chest template model with correct uv mapping.

Fixed enderman model, player models, and piglin model.

Removed the Minecraft Earth Mobs as they will be going in their own plugin to avoid clutter of this one. There are getting too many so i thought it would be best to seperate these two distinct things out.